pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "workflowx"
version = "0.4.0"
description = "Observe workflows. Understand intent. Replace friction. Measure improvement."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Paul Jialiang Wu", email = "wjlgatech@gmail.com" },
]
keywords = ["workflow", "ai", "process-mining", "agents", "automation", "observability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Office/Business",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "structlog>=23.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
capture = [
    "sqlite-utils>=3.35",    # Read Screenpipe SQLite DB
]
inference = [
    "anthropic>=0.40",
    "openai>=1.0",
    "tiktoken>=0.5",
]
replacement = [
    "pyyaml>=6.0",           # Generate Agenticom workflow YAML
    "httpx>=0.27",           # Call Agenticom API
]
api = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]
mcp = [
    "mcp>=1.0",               # MCP server for Claude/Cursor integration
]
all = [
    "workflowx[capture,inference,replacement,api,mcp]",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.10",
    "pre-commit>=3.5",
]

[project.scripts]
workflowx = "workflowx.cli.main:cli"

[project.urls]
Homepage = "https://github.com/wjlgatech/workflowx"
Documentation = "https://github.com/wjlgatech/workflowx/tree/main/docs"
Repository = "https://github.com/wjlgatech/workflowx"
Issues = "https://github.com/wjlgatech/workflowx/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/workflowx"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=workflowx --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
